LayoutTests/media/video-controls-no-scripting.html

<!doctype html>
<html>
    <head>
        <script src="video-test.js" type="text/javascript"></script>
        <script>
        function setup()
        {
            if (!window.layoutTestController) {
                consoleWrite("This test only works in DRT.");
                return;
            }

            video = document.getElementById("fr").contentDocument.querySelector("video");
            testExpected("internals.shadowRoot(video)", null, "!=");
            controls =  internals.shadowRoot(video).firstChild.firstChild;
            testExpected("controls.offsetTop + controls.offsetHeight", video.offsetHeight, "<=");
            testExpected("getComputedStyle(controls).opacity", 1);
            
            consoleWrite("");
            endTest();
        }
        </script>
    </head>

    <body>

        <iframe sandbox src="video-controls-no-scripting-iframe.html" id="fr" style="width: 400px; height: 320px; border: 1px solid black;"></iframe>

        <p>Tests that the built-in controls are always enabled when JavaScript is disabled.</p>
        <pre id="console"></pre>
        <script>
            document.getElementById("fr").onload = setup;
        </script>
    </body>
</html>
