LayoutTests/media/video-controls-no-scripting-expected.txt


Tests that the built-in controls are always enabled when JavaScript is disabled.

EXPECTED (internals.shadowRoot(video) != 'null') OK
EXPECTED (controls.offsetTop + controls.offsetHeight <= '240') OK
EXPECTED (getComputedStyle(controls).opacity == '1') OK

END OF TEST
